enhance :Improved search modal readability using helper functions and hooks
