+++ src/hooks/useDebounce.js
@@ -1,0 +1,25 @@
+import { useState, useEffect } from "react";
+
+/**
+ * Debounce bir değeri belirtilen gecikme süresi sonunda döndüren hook
+ * @param {any} value - Debounce edilecek değer
+ * @param {number} delay - Gecikme süresi (ms)
+ * @returns {any} - Debounce edilmiş değer
+ */
+function useDebounce(value, delay = 500) {
+  const [debouncedValue, setDebouncedValue] = useState(value);
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setDebouncedValue(value);
+    }, delay);
+
+    return () => {
+      clearTimeout(timer);
+    };
+  }, [value, delay]);
+
+  return debouncedValue;
+}
+
+export default useDebounce;
